Added DB Functionality to Sign In
Set up the login page to compare user input passwords to passwords in the database. Once registration is set up with hashed passwords and salts, the login method will need updated.
