+++ iAttendTFL_WebApp/Views/Home/Login.cshtml
@@ -5,7 +5,7 @@
 <div class="border-div small-div">
     <h2 class="hor-center">Log In</h2>
     <br />
-    @using (Html.BeginForm("TestAttemptLogin", "Home", FormMethod.Post))
+    @using (Html.BeginForm("AttemptLogin", "Accounts", FormMethod.Post))
     {
         <table class="no-design-table">
             <tr>
